+++ packages/icons/src/AddIcon.tsx
@@ -1,7 +1,9 @@
 import { shadeBlack, shadeWhite } from '@littlespoon/theme/lib/colors/token'
 import styled from 'styled-components'
 
-export interface OpenIconProps extends React.SVGAttributes<SVGElement> {
+import { getScale } from './utils/css-helpers'
+
+export interface AddIconProps extends React.SVGAttributes<SVGElement> {
   /**
    * The fill of the component. Defaults to white.
    */
@@ -11,20 +13,31 @@ export interface OpenIconProps extends React.SVGAttributes<SVGElement> {
    * The stroke of the component. Defaults to black.
    */
   stroke?: string
+
+  /**
+   * The size of the component. Defaults to "small".
+   */
+  size?: 'xsmall' | 'small' | 'medium' | 'large'
 }
 
-export default function OpenIcon({ fill = shadeWhite, stroke = shadeBlack, ...other }) {
+export default function AddIcon({
+  fill = shadeWhite,
+  stroke = shadeBlack,
+  size = 'small',
+  ...other
+}: AddIconProps): React.ReactElement {
   return (
-    <OpenIconBase
+    <AddIconBase
       width="24"
       height="24"
       viewBox="0 0 24 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
+      size={size}
       {...other}
     >
       <circle cx="12" cy="12" r="12" fill={fill} />
-      <title>Open icon</title>
+      <title>Add icon</title>
       <path
         d="m7.566 7.566 8.868 8.868m-8.868 0 8.868-8.868"
         stroke={stroke}
@@ -32,10 +45,14 @@ export default function OpenIcon({ fill = shadeWhite, stroke = shadeBlack, ...ot
         strokeLinecap="round"
         strokeLinejoin="round"
       />
-    </OpenIconBase>
+    </AddIconBase>
   )
 }
 
-const OpenIconBase = styled.svg<OpenIconProps>`
-  transform: rotate(45deg);
+const AddIconBase = styled.svg<AddIconProps>`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  transition: transform 0.4s;
+  transform: rotate(45deg) scale(${(props) => getScale(props.size)});
 `