+++ packages/icons/src/CalendarIcon.tsx
@@ -1,0 +1,48 @@
+import { shadeBlack } from '@littlespoon/theme/lib/colors/token'
+import styled from 'styled-components'
+
+import { getScale } from './utils/css-helpers'
+
+export interface CalendarIconProps extends React.SVGAttributes<SVGElement> {
+  /**
+   * The fill of the component. Defaults to black.
+   */
+  fill?: string
+
+  /**
+   * The size of the component. Defaults to "small".
+   */
+  size?: 'xsmall' | 'small' | 'medium' | 'large'
+}
+
+export default function CalendarIcon({
+  fill = shadeBlack,
+  size = 'small',
+  ...other
+}: CalendarIconProps): React.ReactElement {
+  return (
+    <CalendarIconBase
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      size={size}
+      {...other}
+    >
+      <title>Calendar icon</title>
+      <desc>Calendar icon</desc>
+      <path
+        d="M4.5 22C4.1 22 3.75 21.85 3.45 21.55C3.15 21.25 3 20.9 3 20.5V5C3 4.6 3.15 4.25 3.45 3.95C3.75 3.65 4.1 3.5 4.5 3.5H6.125V2H7.75V3.5H16.25V2H17.875V3.5H19.5C19.9 3.5 20.25 3.65 20.55 3.95C20.85 4.25 21 4.6 21 5V20.5C21 20.9 20.85 21.25 20.55 21.55C20.25 21.85 19.9 22 19.5 22H4.5ZM4.5 20.5H19.5V9.75H4.5V20.5Z"
+        fill={fill}
+      />
+    </CalendarIconBase>
+  )
+}
+
+export const CalendarIconBase = styled.svg<CalendarIconProps>`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  transition: transform 0.4s;
+  transform: scale(${(props) => getScale(props.size)});
+`