+++ packages/icons/src/CloseIcon.tsx
@@ -1,4 +1,7 @@
 import { shadeBlack, shadeWhite } from '@littlespoon/theme/lib/colors/token'
+import styled from 'styled-components'
+
+import { getScale } from './utils/css-helpers'
 
 export interface CloseIconProps extends React.SVGAttributes<SVGElement> {
   /**
@@ -10,21 +13,31 @@ export interface CloseIconProps extends React.SVGAttributes<SVGElement> {
    * The stroke of the component. Defaults to black.
    */
   stroke?: string
+
+  /**
+   * The size of the component. Defaults to "small".
+   */
+  size?: 'xsmall' | 'small' | 'medium' | 'large'
 }
 
-export default function CloseIcon({ fill = shadeWhite, stroke = shadeBlack, ...other }) {
+export default function CloseIcon({
+  fill = shadeWhite,
+  stroke = shadeBlack,
+  size = 'small',
+  ...other
+}: CloseIconProps): React.ReactElement {
   return (
-    <svg
+    <CloseIconBase
+      xmlns="http://www.w3.org/2000/svg"
       width="24"
       height="24"
       viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
+      size={size}
       {...other}
     >
-      <circle cx="12" cy="12" r="12" fill={fill} />
       <title>Close icon</title>
       <desc>Close icon</desc>
+      <circle cx="12" cy="12" r="12" fill={fill} />
       <path
         d="M7.56567 7.56583L16.4342 16.4343M7.56568 16.4343L16.4342 7.56583"
         stroke={stroke}
@@ -32,6 +45,14 @@ export default function CloseIcon({ fill = shadeWhite, stroke = shadeBlack, ...o
         strokeLinecap="round"
         strokeLinejoin="round"
       />
-    </svg>
+    </CloseIconBase>
   )
 }
+
+export const CloseIconBase = styled.svg<CloseIconProps>`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  transition: transform 0.4s;
+  transform: scale(${(props) => getScale(props.size)});
+`