+++ packages/storybook/src/stories/icons/AccountIcon.stories.tsx
@@ -1,12 +1,12 @@
-import CloseIcon from '@littlespoon/icons/src/CloseIcon'
+import AccountIcon from '@littlespoon/icons/src/AccountIcon'
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 
 export default {
-  title: 'Design System/Icons/CloseIcon',
-  component: CloseIcon,
-} as ComponentMeta<typeof CloseIcon>
+  title: 'Design System/Icons/AccountIcon',
+  component: AccountIcon,
+} as ComponentMeta<typeof AccountIcon>
 
-const Template: ComponentStory<typeof CloseIcon> = (args) => <CloseIcon {...args} />
+const Template: ComponentStory<typeof AccountIcon> = (args) => <AccountIcon {...args} />
 
 export const Default = Template.bind({})
 Default.args = {}