+++ packages/ui/icons/OpenIcon/index.ts
@@ -1,2 +1,0 @@
-export { default } from '@littlespoon/icons/lib/OpenIcon'
-export * from '@littlespoon/icons/lib/OpenIcon'